requirements.txt
fastapi==0.110.0
pytest==8.0.2
httpx==0.27.0
uvicorn==0.29.0
PyJWT==2.8.0
bcrypt==3.2.2
python-multipart==0.0.9
passlib
python-jose